refactor: 게임 config profile을 파일별로 분리
