+++ backend/src/config/game-config-profiles/default.profile.ts
@@ -1,0 +1,3 @@
+import type { GameConfigUpdate } from "../game-config.types";
+
+export const defaultGameConfigProfile: GameConfigUpdate = {};
